requirements.txt
ollama
python-dotenv
tabulate
colorama
requests
fastapi
uvicorn
cloudinary